+++ definitions/global.d.ts
@@ -1,4 +1,3 @@
 declare module "child_process";
 declare module "detect-node";
 declare module "node-fetch-polyfill";
-declare module "camelcase";
