+++ package.json
@@ -15,7 +15,6 @@
   "license": "MIT",
   "dependencies": {
     "@types/uuid": "^3.4.3",
-    "camelcase": "^5.0.0",
     "child_process": "^1.0.2",
     "detect-node": "^2.0.3",
     "node-fetch-polyfill": "^2.0.6",
